+++ docs/CHANGELOG.md
@@ -53,6 +53,8 @@ Version 1.3.0 (Bakeapple): Not yet released
 - Add level editor `Delete` and `Bridge` tools for vertices in vertex mode
 - Add `Mesh_Animate`, `Mesh_Set_Texture`, and `Mesh_Set_Collision` events
 - Add `Reload Meshes` and `Reload Textures` options to level editor `Tools` menu, to scan for new assets in `\user_maps\`
+- Add `To Mesh` tool for converting brushes to `.v3m` static meshes directly in the level editor
+- Improve level editor packfile creation process to include meshes and animations from `Mesh` objects and `Switch_Model`, `Play_Animation`, and `Mesh_Animate` events
 
 [@is-this-c](https://github.com/is-this-c)
 - Replace `os_sleep` in `frametime_calculate` with `wait_for`